+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Hotkeys have been revamped (The previous settings have been reset)
 - Window Cycle hotkey (Only available when Window Monitor is enabled)
+- Sync File Dialog supports Explorer tabs
 
 ## [0.9.6] - 2026-06-19
 - Sync File Dialog with Explorer